requirements.txt
streamlit==1.29.0
tensorflow==2.15.0
streamlit-authenticator==0.1.5

python-dotenv

google-generativeai